add perms f.t i need eval

+++ commands/delete.js
@@ -2,6 +2,8 @@
 const Discord = require('discord.js');
 
 module.exports.run = async (client, message, args) => {
+	const adminperms = new Discord.Permissions('ADMINISTRATOR');
+	if (!message.member.hasPermission(adminperms)) return message.reply('You arent allowed to do that!');
 	if (!args[0]) return message.channel.send('Please give a tag name.');
 	const deleteTag = await client.data.deleteTag(args[0]);
 	message.channel.send(deleteTag);
